requirements.txt
Flask==3.0.3
confluent-kafka==2.5.3
boto3==1.35.24
redis==5.0.8
transformers==4.44.2
torch==2.4.1
requests==2.32.3
python-dotenv==1.0.1